+++ lib/components/date_pickers/components/m3e_date_picker_mode_toggle.dart
@@ -47,7 +47,6 @@ class M3EDatePickerModeToggle extends StatelessWidget {
                   : M3EDatePickerMode.day,
             );
           },
-          haptic: M3EHapticFeedback.none,
           semanticLabel: mode == M3EDatePickerMode.day
               ? localizations.selectYearSemanticsLabel
               : localizations.formatMonthYear(monthDate),