+++ lib/components/toggle_button/m3e_toggle_button.dart
@@ -260,7 +260,7 @@ class M3EToggleButton extends StatefulWidget {
 
   /// M3EHapticFeedback.
   M3EHapticFeedback get decorationHaptic =>
-      decoration?.haptic ?? M3EHapticFeedback.light;
+      decoration?.haptic ?? M3EHapticFeedback.none;
 
   /// double.
   double? get decorationBorderRadius => decoration?.borderRadius;